feat: add combo builder for products
